feat(themable_lottie): Renamed new classes using better names
