+++ src/AddIns/Uno.UI.Lottie/ThemableLottieVisualSource.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿#if !__ANDROID__
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Threading;
@@ -8,61 +9,38 @@ using Windows.UI.Xaml.Data;
 using Uno.Disposables;
 using Windows.UI;
 using Windows.UI.Xaml.Controls;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using Uno.UI.Lottie;
+using System.Linq;
+using System.Json;
 using Uno.Extensions;
-
-// **********************************************************
-// *                        ? WHY ?                         *
-// **********************************************************
-// *                                                        *
-// * This file exists only because there's a bug in the     *
-// * Xamarin Lottie Player as described in this GH issue:   *
-// * https://github.com/Baseflow/LottieXamarin/issues/303   *
-// *                                                        *
-// * Once this bug is resolved, it will be possible to      *
-// * remove this file and the dependency to Netwonsoft.Json *
-// * for the Android target of this project.                *
-// *                                                        *
-// **********************************************************
 
 namespace Microsoft.Toolkit.Uwp.UI.Lottie
 {
 	[Bindable]
-	public class DynamicReloadedLottieAnimatedVisualSource : LottieVisualSourceBase, IDynamicAnimatedVisualSource
+	public partial class ThemableLottieVisualSource : LottieVisualSourceBase, IThemableAnimatedVisualSource
 	{
-		private JObject? _currentDocument;
-
-		private readonly Dictionary<string, ColorBinding> _colorsBindings
-			= new Dictionary<string, ColorBinding>(2);
+		private JsonValue? _currentDocument;
 
 		private UpdatedAnimation? _updateCallback;
 		private string? _sourceCacheKey;
 
-		public void SetColorProperty(string propertyName, Color? color)
+		protected override bool IsPayloadNeedsToBeUpdated => true;
+
+#if NETFRAMEWORK
+		public Task LoadForTests(
+			IInputStream sourceJson,
+			string sourceCacheKey,
+			UpdatedAnimation updateCallback)
 		{
-			if (_colorsBindings.TryGetValue(propertyName, out var existing))
-			{
-				existing.NextValue = color;
-			}
-			else
-			{
-				_colorsBindings[propertyName] = new ColorBinding { NextValue = color };
-			}
-
-			if (_currentDocument == null)
-			{
-				return; // no document to change yet
-			}
-
-			if (ApplyProperties())
-			{
-				NotifyCallback();
-			}
+			_updateCallback = updateCallback;
+			return LoadAndUpdate(default, sourceCacheKey, sourceJson);
 		}
 
-		protected override bool IsPayloadNeedsToBeUpdated => true;
+		public string? GetJson()
+		{
+			return _currentDocument?.ToString();
+		}
+#endif
 
 		protected override IDisposable? LoadAndObserveAnimationData(
 			IInputStream sourceJson,
@@ -93,17 +71,13 @@ namespace Microsoft.Toolkit.Uwp.UI.Lottie
 			// because System.Text.Json does not support changing the
 			// parsed document - it's read only.
 
-			// LOAD JSON
-			JObject document;
-			using (var stream = sourceJson.AsStreamForRead(0))
-			{
-				using var streamReader = new StreamReader(stream);
-				using var reader = new JsonTextReader(streamReader);
-				document = JObject.Load(reader);
-			}
+			// LOAD & PARSE JSON
+			LoadAndParseDocument(sourceJson);
 
-			// PARSE JSON
-			ParseDocument(document);
+			if(_currentDocument == null)
+			{
+				return;
+			}
 
 			// APPLY PROPERTIES
 			ApplyProperties();
@@ -112,35 +86,48 @@ namespace Microsoft.Toolkit.Uwp.UI.Lottie
 			NotifyCallback();
 		}
 
-		private void ParseDocument(JObject document)
+		private void LoadAndParseDocument(IInputStream sourceJson)
 		{
+			JsonObject? document;
+			using (var stream = sourceJson.AsStreamForRead(0))
+			{
+				document = JsonValue.Load(stream) as JsonObject;
+			}
+
+			if (document == null)
+			{
+				return;
+			}
+
 			_currentDocument = document;
 
 			foreach (var colorBinding in _colorsBindings)
 			{
 				colorBinding.Value.Elements.Clear();
+				colorBinding.Value.NextValue = colorBinding.Value.CurrentValue;
 			}
 
-			void ParseLayers(JToken layersElement)
+			void ParseLayers(JsonArray layers)
 			{
-				if (!(layersElement is JArray layers))
+				if (layers == null)
 				{
 					return; // potentially invalid lottie file
 				}
 
 				foreach (var layer in layers)
 				{
-					if (layer is JObject l)
+					if (layer is JsonObject l)
 					{
-						var shapesValue = l.GetValue("shapes");
-
-						if (shapesValue is JArray shapes)
+						if (l.TryGetValue("shapes", out var shapesValue))
 						{
-							foreach (var shape in shapes)
+							if (shapesValue is JsonArray shapes)
 							{
-								if (shape is JObject s)
+								foreach (var shape in shapes)
 								{
-									ParseShape(s);
+									if (shape is JsonObject s)
+									{
+										ParseShape(s);
+									}
 								}
 							}
 						}
@@ -148,27 +135,33 @@ namespace Microsoft.Toolkit.Uwp.UI.Lottie
 				}
 			}
 
-			void ParseShape(JObject shapeElement)
+			void ParseShape(JsonObject shapeElement)
 			{
-				var typeValue = shapeElement.GetValue("ty");
-				if (typeValue.Type != JTokenType.String)
+				if (!shapeElement.TryGetValue("ty", out var typeValue))
+				{
+					return; // potentially invalid lottie file
+				}
+				if (typeValue.JsonType != JsonType.String)
 				{
 					return; // potentially invalid lottie file
 				}
 
-				var shapeType = typeValue.Value<string>();
+				var shapeType = (string)typeValue;
 
 				if (shapeType != null && shapeType.Equals("gr"))
 				{
 					// That's a group
 
-					var itemsProperty = shapeElement.GetValue("it");
-
-					if (itemsProperty is JArray items)
+					if (!shapeElement.TryGetValue("it", out var itemsProperty)
+						|| itemsProperty.JsonType != JsonType.Array)
+					{
+						return; // potentially invalid lottie file
+					}
+					if (itemsProperty is JsonArray items)
 					{
 						foreach (var item in items)
 						{
-							if (item is JObject s)
+							if (item is JsonObject s)
 							{
 								ParseShape(s);
 							}
@@ -178,14 +171,13 @@ namespace Microsoft.Toolkit.Uwp.UI.Lottie
 					return;
 				}
 
-				var nameProperty = shapeElement.GetValue("nm");
-
-				if (nameProperty.Type != JTokenType.String)
+				if(!shapeElement.TryGetValue("nm", out var nameProperty)
+					|| nameProperty.JsonType != JsonType.String)
 				{
 					return; // No name
 				}
 
-				var name = nameProperty.Value<string>();
+				var name = (string)nameProperty;
 
 				if (!string.IsNullOrWhiteSpace(name))
 				{
@@ -212,9 +204,9 @@ namespace Microsoft.Toolkit.Uwp.UI.Lottie
 				}
 			}
 
-			if (document.TryGetValue("layers", out var layers))
+			if (document.TryGetValue("layers", out var lyrs) && lyrs is JsonArray documentLayers)
 			{
-				ParseLayers(layers);
+				ParseLayers(documentLayers);
 			}
 		}
 
@@ -223,28 +215,31 @@ namespace Microsoft.Toolkit.Uwp.UI.Lottie
 			var changed = false;
 			foreach (var colorBinding in _colorsBindings)
 			{
-				if (!(colorBinding.Value.NextValue is { } color))
+				if (!(colorBinding.Value.NextValue is {} color))
 				{
 					continue; // nothing to change
 				}
 
-				var colorComponents = new[] { color.R / 255f, color.G / 255f, color.B / 255f, color.A / 255f };
+				var colorComponents = new[] {color.R / 255f, color.G / 255f, color.B / 255f, color.A / 255f};
 
 				foreach (var element in colorBinding.Value.Elements)
 				{
-					var k = (element.GetValue("c") as JObject)?.GetValue("k") as JArray;
-
-					if (k != null)
+					if (element.TryGetValue("c", out var cElm)
+						&& cElm is JsonObject c
+						&& c.TryGetValue("k", out var kElm)
+						&& kElm is JsonArray k)
 					{
+
 						k.Clear();
-						k.Add(new JValue(colorComponents[0]));
-						k.Add(new JValue(colorComponents[1]));
-						k.Add(new JValue(colorComponents[2]));
-						k.Add(new JValue(colorComponents[3]));
+						k.Add(colorComponents[0]);
+						k.Add(colorComponents[1]);
+						k.Add(colorComponents[2]);
+						k.Add(colorComponents[3]);
 
 						changed = true;
 					}
 				}
+
 				colorBinding.Value.CurrentValue = colorBinding.Value.NextValue;
 				colorBinding.Value.NextValue = null;
 			}
@@ -254,9 +249,9 @@ namespace Microsoft.Toolkit.Uwp.UI.Lottie
 
 		private void NotifyCallback()
 		{
-			if (_updateCallback is { } callback)
+			if (_updateCallback is {} callback)
 			{
-				var json = _currentDocument?.ToString(Formatting.None);
+				var json = _currentDocument?.ToString();
 				if (json is { })
 				{
 					var propertiesKey = _colorsBindings
@@ -271,9 +266,10 @@ namespace Microsoft.Toolkit.Uwp.UI.Lottie
 
 		private class ColorBinding
 		{
-			internal List<JObject> Elements { get; } = new List<JObject>(1);
+			internal List<JsonObject> Elements { get; } = new List<JsonObject>(1);
 			internal Color? CurrentValue { get; set; }
 			internal Color? NextValue { get; set; }
 		}
 	}
 }
+#endif
