--- src/Uno.UI/UI/Xaml/Controls/AnimatedVisualPlayer/IDynamicAnimatedVisualSource.cs
@@ -1,7 +1,0 @@
-namespace Windows.UI.Xaml.Controls
-{
-	public partial interface IDynamicAnimatedVisualSource : IAnimatedVisualSource
-	{
-		void SetColorProperty(string propertyName, Windows.UI.Color? color);
-	}
-}
